refactor the modbus connection code structure
