+++ src/main/java/org/edgexfoundry/domain/ModbusValueType.java
@@ -1,7 +1,7 @@
 package org.edgexfoundry.domain;
 
 public enum ModbusValueType {
-    FLOAT32(2), FLOAT64(4), INT16(1), INT32(2), INT64(4) , BOOLEAN(1) ;
+    FLOAT32(2), FLOAT64(4), INT16(1), INT32(2), INT64(4), BOOLEAN(1);
 
     int length;
